requirements.txt
mysql-connector-python==8.3.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySide6==6.7.0
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
shiboken6==6.7.0
